# D-Voyager-BackEnd/shuttle-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shuttle System API",
    "description": "Dokumentasi resmi RESTful API untuk Sistem Pemesanan Shuttle (Travel). Dibuat untuk memenuhi kriteria Final Project.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/D-Voyager-BackEnd/public",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": [
    { "name": "Auth", "description": "Otentikasi Pengguna" },
    { "name": "Public", "description": "Informasi Publik (Tanpa Login)" },
    { "name": "Customer", "description": "Fitur Khusus Penumpang" },
    { "name": "Driver", "description": "Fitur Khusus Supir" }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login User",
        "description": "Mendapatkan Bearer Token untuk akses API.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "customer1@shuttle.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string", "example": "1|LaravelSanctumToken..." },
                    "user": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/schedules": {
      "get": {
        "tags": ["Public"],
        "summary": "Lihat Jadwal Tersedia",
        "description": "Menampilkan daftar jadwal keberangkatan shuttle.",
        "responses": {
          "200": {
            "description": "Berhasil mengambil data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "object" }
                }
              }
            }
          }
        }
      }
    },
    "/api/customer/my-bookings": {
      "get": {
        "tags": ["Customer"],
        "summary": "Riwayat Pesanan Customer",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List Pesanan",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "type": "object" } }
              }
            }
          }
        }
      }
    },
    "/api/customer/booking": {
      "post": {
        "tags": ["Customer"],
        "summary": "Buat Pesanan Baru (Booking)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "schedule_id": { "type": "integer", "example": 1 },
                  "seats": {
                    "type": "array",
                    "items": { "type": "string" },
                    "example": ["1", "2"]
                  }
                },
                "required": ["schedule_id", "seats"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking Sukses"
          }
        }
      }
    },
    "/api/driver/my-schedules": {
      "get": {
        "tags": ["Driver"],
        "summary": "Lihat Jadwal Tugas Supir",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List Jadwal"
          }
        }
      }
    },
    "/api/driver/schedule/{id}/start": {
      "post": {
        "tags": ["Driver"],
        "summary": "Mulai Perjalanan (Start Trip)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID Jadwal"
          }
        ],
        "responses": {
          "200": {
            "description": "Perjalanan Dimulai"
          }
        }
      }
    }
  }
}